refactor : 알림 설정 폼·데이터 레이어를 mypage _shared로 승격
모달과 모바일 설정 페이지가 공유하는 폼·훅·스키마를 프렉탈 경계 규칙에 맞게 _shared(ui·api·model)로 이동
