+++ apps/web/src/app/partner/mypage/_model/types.ts
@@ -7,10 +7,6 @@ import type {
   mypageCertificationSchema,
   mypageSchema,
 } from "./mypage.schema";
-import type {
-  notificationSettingsSchema,
-  updateNotificationSettingsBodySchema,
-} from "./notification-settings.schema";
 
 export type UserRole = z.infer<typeof userRoleSchema>;
 export type MypageProfile = z.infer<typeof mypageProfileSchema>;
@@ -18,7 +14,3 @@ export type MypageStats = z.infer<typeof mypageStatsSchema>;
 export type MypageMonthlyActivity = z.infer<typeof mypageMonthlyActivitySchema>;
 export type MypageCertification = z.infer<typeof mypageCertificationSchema>;
 export type Mypage = z.infer<typeof mypageSchema>;
-export type NotificationSettings = z.infer<typeof notificationSettingsSchema>;
-export type UpdateNotificationSettingsBody = z.infer<
-  typeof updateNotificationSettingsBodySchema
->;
