+++ apps/web/src/app/partner/mypage/_shared/api/use-notification-settings.ts
