+++ apps/web/src/app/partner/mypage/_shared/model/notification-settings.schema.ts
